+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
 ### Fixed
 - Script execution on java > 11 ([#12](https://github.com/scm-manager/scm-script-plugin/pull/12)) 
+- Loading of plugin classes in a script ([#13](https://github.com/scm-manager/scm-script-plugin/pull/13))
 
 ### Added
 - Documentation in English and German ([#8](https://github.com/scm-manager/scm-script-plugin/pull/8))